Queue the definition of jobs upfront, not the jobs themselves
Creating the render job forces an upfront creation of an image
to store the renderer map immediately. So storing many thousands
of tasks is also storing all the images for these in memory
upfront, leading to crashes.

Instead just queue the map settings for each job, and create the
actual task when we are ready to start rendering that particular
job.

Fixes crashes when memory runs out.